chore: Remove last mutation (hiding tracks)
